remove all leftovers
