--- src/features/shields2/index.ts
@@ -1,53 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this file,
- * You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-export {
-  ShieldsPanel,
-  ShieldsHeader,
-  MainToggle,
-  SiteOverview,
-  TotalBlockedStats,
-  SiteInfo,
-  //
-  BlockedInfoRow,
-  BlockedInfoRowSingle,
-  BlockedInfoRowForSelect,
-  BlockedInfoRowData,
-  BlockedInfoRowDataForSelect,
-  // footer
-  MainFooter,
-  //
-  BlockedListHeader,
-  BlockedListContent,
-  BlockedListSummary,
-  BlockedListStatic,
-  BlockedListDynamic,
-  BlockedListItemHeader,
-  BlockedListItem,
-  BlockedListItemWithOptions,
-  BlockedListFooter,
-  BlockedListFooterWithOptions,
-  DisabledContentView
-} from './structure'
-
-export {
-  MainToggleHeading,
-  MainToggleText,
-  ToggleStateText,
-  Favicon,
-  SiteInfoText,
-  TotalBlockedStatsNumber,
-  TotalBlockedStatsText,
-  BlockedInfoRowStats,
-  BlockedInfoRowText,
-  BlockedListSummaryText,
-  BlockedListItemHeaderStats,
-  BlockedListItemHeaderText,
-  Link,
-  LinkAction,
-  DisabledContentText,
-  ShieldsButton
-} from './display'
-
-export { ArrowDownIcon, ArrowUpIcon, ShieldIcon } from './media'
