+++ stories/features/shields/components/interfaceControls.tsx
@@ -4,124 +4,50 @@
 
 import * as React from 'react'
 
-// Feature-specific components
-import {
-  ShowMoreIcon,
-  Toggle,
-  StatFlex,
-  ResourcesSwitchLabel,
-  EmptyButton,
-  ToggleFlex,
-  ToggleGrid
-} from '../../../../src/features/shields'
-
-// Component groups
-import BlockedResources from './blockedReources/blockedResources'
-import StaticList from './blockedReources/staticList'
-
-// Fake data
-import locale from '../fakeLocale'
-import data from '../fakeData'
-
-const totalAdsTrackersBlockedList = (favicon: string, sitename: string, onToggle: () => void) => {
-  return (
-    <BlockedResources
-      favicon={favicon}
-      sitename={sitename}
-      title={locale.blockAds}
-      onToggle={onToggle}
-      data={data.totalAdsTrackersBlocked}
-    >
-      <StaticList onClickDismiss={onToggle} list={data.blockedFakeResources} />
-    </BlockedResources>
-  )
-}
-
-const connectionsEncryptedList = (favicon: string, sitename: string, onToggle: () => void) => {
-  return (
-    <BlockedResources
-      favicon={favicon}
-      sitename={sitename}
-      title={locale.connectionsEncrypted}
-      onToggle={onToggle}
-      data={data.thirdPartyDeviceRecognitionBlocked}
-    >
-      <StaticList onClickDismiss={onToggle} list={data.blockedFakeResources} />
-    </BlockedResources>
-  )
-}
+// Group Components
+import AdsTrackersControl from './controls/adsTrackersControl'
+import HTTPSUpgradesControl from './controls/httpsUpgradesControl'
 
 interface Props {
-  enabled: boolean
-  sitename: string
   favicon: string
+  hostname: string
+  isBlockedListOpen: boolean
+  setBlockedListOpen: () => void
+  adsTrackersBlocked: number
+  httpsUpgrades: number
 }
 
 interface State {
-  blockAds: boolean
-  connectionsEncrypted: boolean
-  openTotalAdsTrackersBlockedList: boolean
-  openConnectionsEncryptedList: boolean
+  connectionsUpgradedOpen: boolean
+  connectionsUpgradedEnabled: boolean
 }
 
 export default class InterfaceControls extends React.PureComponent<Props, State> {
-  constructor (props: Props) {
-    super(props)
-    this.state = {
-      blockAds: false,
-      openTotalAdsTrackersBlockedList: false,
-      connectionsEncrypted: false,
-      openConnectionsEncryptedList: false
-    }
-  }
-
-  onChangeBlockAds = (event: React.ChangeEvent<HTMLInputElement>) => {
-    this.setState({ blockAds: event.target.checked })
-  }
-
-  onChangeConnectionsEncrypted = (event: React.ChangeEvent<HTMLInputElement>) => {
-    this.setState({ connectionsEncrypted: event.target.checked })
-  }
-
-  onToggleTotalAdsTrackersBlocked = () => {
-    this.setState({ openTotalAdsTrackersBlockedList: !this.state.openTotalAdsTrackersBlockedList })
-  }
-
-  onToggleConnectionsEncrypted = () => {
-    this.setState({ openConnectionsEncryptedList: !this.state.openConnectionsEncryptedList })
-  }
-
   render () {
-    const { enabled, favicon, sitename } = this.props
     const {
-      blockAds,
-      openTotalAdsTrackersBlockedList,
-      connectionsEncrypted,
-      openConnectionsEncryptedList
-    } = this.state
-
-    if (!enabled) {
-      return null
-    }
-
+      favicon,
+      hostname,
+      isBlockedListOpen,
+      setBlockedListOpen,
+      adsTrackersBlocked,
+      httpsUpgrades
+    } = this.props
     return (
       <>
-        {/* ads toggle */}
-          <ToggleGrid>
-            <EmptyButton onClick={this.onToggleTotalAdsTrackersBlocked}><ShowMoreIcon /></EmptyButton>
-            <StatFlex onClick={this.onToggleTotalAdsTrackersBlocked}>{data.totalAdsTrackersBlocked}</StatFlex>
-            <ResourcesSwitchLabel onClick={this.onToggleTotalAdsTrackersBlocked}>{locale.blockAds}</ResourcesSwitchLabel>
-            <ToggleFlex><Toggle id='blockAds' checked={blockAds} onChange={this.onChangeBlockAds} /></ToggleFlex>
-          </ToggleGrid>
-          {openTotalAdsTrackersBlockedList ? totalAdsTrackersBlockedList(favicon, sitename, this.onToggleTotalAdsTrackersBlocked) : null}
-        {/* connections encrypted toggle */}
-        <ToggleGrid>
-          <EmptyButton disabled={true} onClick={this.onToggleConnectionsEncrypted}><ShowMoreIcon /></EmptyButton>
-          <StatFlex disabled={true} onClick={this.onToggleConnectionsEncrypted}>0</StatFlex>
-          <ResourcesSwitchLabel disabled={true} onClick={this.onToggleConnectionsEncrypted}>{locale.connectionsEncrypted}</ResourcesSwitchLabel>
-          <ToggleFlex><Toggle id='connectionsEncrypted' checked={connectionsEncrypted} onChange={this.onChangeConnectionsEncrypted} /></ToggleFlex>
-        </ToggleGrid>
-        {openConnectionsEncryptedList ? connectionsEncryptedList(favicon, sitename, this.onToggleConnectionsEncrypted) : null}
+        <AdsTrackersControl
+          favicon={favicon}
+          hostname={hostname}
+          isBlockedListOpen={isBlockedListOpen}
+          adsTrackersBlocked={adsTrackersBlocked}
+          setBlockedListOpen={setBlockedListOpen}
+        />
+        <HTTPSUpgradesControl
+          favicon={favicon}
+          hostname={hostname}
+          isBlockedListOpen={isBlockedListOpen}
+          httpsUpgrades={httpsUpgrades}
+          setBlockedListOpen={setBlockedListOpen}
+        />
       </>
     )
   }
