+++ stories/features/shields/components/list/static.tsx
@@ -17,7 +17,7 @@ import {
   BlockedInfoRowStats,
   BlockedListSummaryText,
   ShieldsButton
-} from '../../../../../src/features/shields2'
+} from '../../../../../src/features/shields'
 
 // Fake data
 import { getLocale } from '../../fakeLocale'
